Support older .sd2 files (#14)
* Update to FauFau 1.4.0

* Update fields.txt

Adding names used in older versions

* Misc fixes and enhancements

Fixes for the Get Hash and Decrypt buttons. Remove some unused things and created DebugStringHashCoverage function to print info about whether we are missing hashes or not. Add support for loading test-fields.txt so that we can differentiate known good strings from things we want to test. Savable strings map will be populated with good strings and any strings that match within the given db, which can be used (but not implemented as a button yet) to write a new fields.txt.

+++ SDBrowser/PgDB/PGImportExport.cs
@@ -136,7 +136,7 @@ namespace SDBrowser
             var dbMeta = new SdbMeta
             {
                 Patch        = DB.Patch,
-                Flags        = DB.Flags,
+                Flags        = (uint)DB.Flags,
                 Timestamp    = DB.Timestamp,
                 FileVersion  = (uint)GetInstanceField(DB.GetType(), DB, "fileVersion"),
                 TableVersion = (uint)GetInstanceField(DB.GetType(), DB, "tableVersion")
